feat: Seperate ImageUploader component
