+++ React/panda-market/src/pages/RegisterItemPage/RegisterItemPage.jsx
@@ -1,5 +1,5 @@
-import { useState, useEffect, useRef } from "react";
-import PrusIcon from "../../assets/icon/ic_plus.svg";
+import { useState, useEffect } from "react";
+import ImageUploader from "./components/ImageUploader";
 import DeleteIcon from "../../assets/icon/ic_delete.svg";
 import styles from "./RegisterItemPage.module.css";
 
@@ -12,39 +12,9 @@ function RegisterItemPage() {
     tag: "",
   });
   const [tagValues, setTagValues] = useState([]);
-  const [itemImg, setItemImg] = useState(null);
-  const [fileError, setFileError] = useState("");
-  const fileInputRef = useRef(null);
 
   const handleRegister = async (event) => {
     event.preventDefault();
-  };
-
-  const handleFileButtonClick = () => {
-    if (!fileInputRef.current) return;
-    fileInputRef.current.click();
-  };
-
-  const handleFileChange = (event) => {
-    const files = event.target.files;
-
-    if (!files) {
-      return;
-    }
-
-    if (files.length > 1) {
-      setFileError("*이미지 등록은 최대 1개까지 가능합니다.");
-      return;
-    }
-
-    setFileError("");
-    setItemImg(URL.createObjectURL(files[0]));
-  };
-
-  const handleDeleteFile = () => {
-    setItemImg(null);
-    setFileError("");
-    fileInputRef.current.value = "";
   };
 
   const handleChange = (event) => {
@@ -93,42 +63,7 @@ function RegisterItemPage() {
       <div className={styles.itemFormSection}>
         <div className={styles.infoInputSection}>
           <h2 className={styles.title}>상품 이미지</h2>
-          <div className={styles.imgSectionContainer}>
-            <button
-              type="button"
-              className={styles.imgAddButton}
-              onClick={handleFileButtonClick}
-            >
-              <input
-                type="file"
-                ref={fileInputRef}
-                style={{ display: "none" }}
-                onChange={handleFileChange}
-                accept=".jpeg, .jpg, .png"
-                multiple
-              />
-              <div className={styles.imgAddButtonContent}>
-                <div>
-                  <img src={PrusIcon} alt="" />
-                </div>
-                <div className={styles.imgAddButtonText}>이미지 등록</div>
-              </div>
-            </button>
-            {itemImg && (
-              <div className={styles.imgPreview}>
-                <img src={itemImg} alt="상품 이미지 미리보기" />
-                <button type="button">
-                  <img
-                    src={DeleteIcon}
-                    alt=""
-                    className={styles.deleteButton}
-                    onClick={handleDeleteFile}
-                  />
-                </button>
-              </div>
-            )}
-          </div>
-          {fileError && <div className={styles.errorMessage}>{fileError}</div>}
+          <ImageUploader />
         </div>
         <div className={styles.infoInputSection}>
           <h2 className={styles.title}>상품명</h2>
