+++ React/panda-market/src/pages/RegisterItemPage/components/ImageUploader.module.css
@@ -1,0 +1,93 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 1.6rem;
+}
+
+.imgSectionContainer {
+  display: flex;
+  gap: 2.4rem;
+}
+
+.imgAddButton,
+.imgAddButton:hover,
+.imgAddButton:focus,
+.imgAddButton:disabled {
+  width: 28.2rem;
+  height: 28.2rem;
+  border-radius: 1.2rem;
+  background-color: var(--gray100);
+}
+
+.imgAddButtonContent {
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+}
+
+.imgAddButtonText {
+  font-weight: 400;
+  font-size: 1.6rem;
+  color: var(--gray400);
+}
+
+.imgPreview {
+  position: relative;
+  width: 28.2rem;
+  height: 28.2rem;
+}
+
+.imgPreview > img {
+  width: 100%;
+  border-radius: 1.2rem;
+  aspect-ratio: 1 / 1;
+  object-fit: cover;
+}
+
+.imgPreview > button {
+  all: unset;
+  cursor: pointer;
+  position: absolute;
+  top: 1.2rem;
+  right: 1.2rem;
+}
+
+.errorMessage {
+  font-weight: 400;
+  font-size: 1.6rem;
+  color: var(--red);
+}
+
+/* Tablet */
+@media (max-width: 1199px) {
+  .imgSectionContainer {
+    gap: 1rem;
+  }
+
+  .imgAddButton {
+    width: 16.8rem;
+    height: 16.8rem;
+  }
+
+  .imgPreview {
+    width: 16.8rem;
+    height: 16.8rem;
+  }
+}
+
+/* Mobile */
+@media (max-width: 767px) {
+  .imgSectionContainer {
+    gap: 1rem;
+  }
+
+  .imgAddButton {
+    width: 16.8rem;
+    height: 16.8rem;
+  }
+
+  .imgPreview {
+    width: 16.8rem;
+    height: 16.8rem;
+  }
+}
